fluid'd cmd.Add, moved SetupMagic, added new wear
